(GP-2.3): Creates BLOB files from source file contents
This commit implements the creation of BLOB files. It computes the SHA-1 hash of the source file's content and stores an exact copy of the source file's bytes in an object file within the `git/objects/` directory, named after the SHA-1 hash. It also adds documentation to the README.

It also ensures that the `git/objects/` directory is created if it doesn't exist and avoids recreating BLOB files if they already exist.
Compression of BLOB files is supported via a toggle.